Added tracking of unopened chests.
This postpones the need for better inventory management code.

If a chest remains unopened for 10 seconds, its id is logged to file. This allows it to survive client reloads and worse (without having to fiddle with local storage).
The openLoggedChests function takes the contents of such a file (easily pasted between backticks) and opens these logged chests over time. It pauses when the character has insufficient inventory (well, probably - there doesn't seem to be a way to get chest contents from id if the chest is no longer loaded in the client) and stops once done with all logged chests.

Chest appear to not be "garbage collected" by the goblin in Wizard's Crib unless the character is offline for a while. This makes it quite safe (loot-wise) to have your characters running unsupervised for long periods of time.