Make PARENT available everywhere.
parent is a reference to window.parent, both of which will become undefined when the user code stops.
This is a problem for code functions (e.g. callbacks) that should continue working even afterwards.
The variable created by `let PARENT = parent;` is not affected by this and should be used instead.